Add initial Docker setup and environment configuration

+++ src/main/java/com/project/quanlycanghangkhong/dto/UserFlightShiftResponseDTO.java
@@ -3,7 +3,7 @@ package com.project.quanlycanghangkhong.dto;
 import java.time.LocalDate;
 
 public class UserFlightShiftResponseDTO {
-    private Long id;
+    private Integer id;
     private Integer userId;
     private String userName;
     private Long flightId;
@@ -11,9 +11,11 @@ public class UserFlightShiftResponseDTO {
     private LocalDate shiftDate;
 
     // Constructors
-    public UserFlightShiftResponseDTO() {}
+    public UserFlightShiftResponseDTO() {
+    }
 
-    public UserFlightShiftResponseDTO(Long id, Integer userId, String userName, Long flightId, String flightNumber, LocalDate shiftDate) {
+    public UserFlightShiftResponseDTO(Integer id, Integer userId, String userName, Long flightId,
+            String flightNumber, LocalDate shiftDate) {
         this.id = id;
         this.userId = userId;
         this.userName = userName;
@@ -21,44 +23,52 @@ public class UserFlightShiftResponseDTO {
         this.flightNumber = flightNumber;
         this.shiftDate = shiftDate;
     }
-    
-    
 
-   
-	// Getters & Setters
-    public Long getId() {
+    // Getters & Setters
+    public Integer getId() {
         return id;
     }
-    public void setId(Long id) {
+
+    public void setId(Integer id) {
         this.id = id;
     }
+
     public Integer getUserId() {
         return userId;
     }
+
     public void setUserId(Integer userId) {
         this.userId = userId;
     }
+
     public String getUserName() {
         return userName;
     }
+
     public void setUserName(String userName) {
         this.userName = userName;
     }
+
     public Long getFlightId() {
         return flightId;
     }
+
     public void setFlightId(Long flightId) {
         this.flightId = flightId;
     }
+
     public String getFlightNumber() {
         return flightNumber;
     }
+
     public void setFlightNumber(String flightNumber) {
         this.flightNumber = flightNumber;
     }
+
     public LocalDate getShiftDate() {
         return shiftDate;
     }
+
     public void setShiftDate(LocalDate shiftDate) {
         this.shiftDate = shiftDate;
     }
