+++ src/main/java/com/project/quanlycanghangkhong/dto/UserFlightShiftResponseSearchDTO.java
@@ -4,23 +4,17 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 
 public class UserFlightShiftResponseSearchDTO {
-    private Long scheduleId;
+    private Integer scheduleId;
     private String userName;
     private String teamName;
     private String unitName;
     private LocalDate shiftDate;
     private String flightNumber;
-    private LocalTime departureTime; // thay thế startTime
-    private LocalTime arrivalTime;   // thay thế endTime
-//    private String location;
-//    private String description;
+    private LocalTime departureTime;
+    private LocalTime arrivalTime;
 
-    // Constructor đầy đủ theo thứ tự của query
-    public UserFlightShiftResponseSearchDTO(Long scheduleId, String userName, String teamName, String unitName,
-                                      LocalDate shiftDate, String flightNumber, LocalTime departureTime
-                                      , LocalTime arrivalTime
-//                                      String location, String description
-                                      ) {
+    public UserFlightShiftResponseSearchDTO(Integer scheduleId, String userName, String teamName, String unitName,
+            LocalDate shiftDate, String flightNumber, LocalTime departureTime, LocalTime arrivalTime) {
         this.scheduleId = scheduleId;
         this.userName = userName;
         this.teamName = teamName;
@@ -29,74 +23,70 @@ public class UserFlightShiftResponseSearchDTO {
         this.flightNumber = flightNumber;
         this.departureTime = departureTime;
         this.arrivalTime = arrivalTime;
-//        this.location = location;
-//        this.description = description;
     }
 
     // Getters & Setters
-    public Long getScheduleId() {
+    public Integer getScheduleId() {
         return scheduleId;
     }
-    public void setScheduleId(Long scheduleId) {
+
+    public void setScheduleId(Integer scheduleId) {
         this.scheduleId = scheduleId;
     }
+
     public String getUserName() {
         return userName;
     }
+
     public void setUserName(String userName) {
         this.userName = userName;
     }
+
     public String getTeamName() {
         return teamName;
     }
+
     public void setTeamName(String teamName) {
         this.teamName = teamName;
     }
+
     public String getUnitName() {
         return unitName;
     }
+
     public void setUnitName(String unitName) {
         this.unitName = unitName;
     }
+
     public LocalDate getShiftDate() {
         return shiftDate;
     }
+
     public void setShiftDate(LocalDate shiftDate) {
         this.shiftDate = shiftDate;
     }
+
     public String getFlightNumber() {
         return flightNumber;
     }
+
     public void setFlightNumber(String flightNumber) {
         this.flightNumber = flightNumber;
     }
-  
+
     public LocalTime getDepartureTime() {
-		return departureTime;
-	}
+        return departureTime;
+    }
 
-	public void setDepartureTime(LocalTime departureTime) {
-		this.departureTime = departureTime;
-	}
+    public void setDepartureTime(LocalTime departureTime) {
+        this.departureTime = departureTime;
+    }
 
-	public LocalTime getArrivalTime() {
-		return arrivalTime;
-	}
+    public LocalTime getArrivalTime() {
+        return arrivalTime;
+    }
 
-	public void setArrivalTime(LocalTime arrivalTime) {
-		this.arrivalTime = arrivalTime;
-	}
-
-//	public String getLocation() {
-//        return location;
-//    }
-//    public void setLocation(String location) {
-//        this.location = location;
-//    }
-//    public String getDescription() {
-//        return description;
-//    }
-//    public void setDescription(String description) {
-//        this.description = description;
-//    }
+    public void setArrivalTime(LocalTime arrivalTime) {
+        this.arrivalTime = arrivalTime;
+    }
 }
