+++ src/main/java/com/project/quanlycanghangkhong/repository/UnitRepository.java
@@ -8,7 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.project.quanlycanghangkhong.model.Unit;
 
 @Repository
-public interface UnitRepository extends JpaRepository<Unit, Long> {
+public interface UnitRepository extends JpaRepository<Unit, Integer> {
     List<Unit> findByTeam_Id(Integer teamId);
-
 }
